+++ frontend/src/app.css
@@ -1,0 +1,10 @@
+html, body {
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 2px;
+}
+
+* {
+  box-sizing: border-box;
+}
