+++ frontend/src/lib/components/GameTopBar.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+	import type { RestrictedGameState } from '$lib/types/GameState';
+	import GameInfo from './GameInfo.svelte';
+
+	export let restrictedGameState: RestrictedGameState;
+</script>
+
+<GameInfo {restrictedGameState}></GameInfo>
